requirements.txt
requests>=2.31.0
psutil>=5.9.0
PyYAML>=6.0
Jinja2>=3.1.0
jsonschema>=4.21.0
grpcio>=1.62.0
protobuf>=4.25.0
pytest>=8.0.0
